refactor(web): extract chart components into src/components/charts/ (issue #642, Phase 3 seam 3 pt 2)
Moves FarFieldChart, SmithChart, CurrentCanvas (+drawArmEnvelope) out
of App.tsx into their own files under components/charts/, along with
their supporting machinery:
 - charts/cuts.ts: the pattern-cuts cache + /ws cuts transport
 - charts/palette.ts: the canvas color/theme helpers
 - charts/types.ts: FarFieldCut/PatternMetrics/PinnedPattern, shared
   between the charts and App.tsx's PatternCompareTable/pin state
 - components/hooks.ts: Theme/ThemeContext, factored out to a leaf
   module so App.tsx and the charts can both use the context value
   without a runtime import cycle (App.tsx imports the charts)

One non-verbatim accommodation: cutsWsSend was a bare `let` mutated
directly by App.tsx's websocket effect. ESM live bindings only allow
an importer to *read* an exported `let`, not assign to it, so
App.tsx's two assignment sites now go through a new `setCutsWsSend`
setter instead. Everything else is a verbatim move (export added,
PatternData exported for FarFieldChart's type-only import).
