+++ src/antennaknobs/web/frontend/src/App.tsx
@@ -39,11 +39,9 @@ import type {
   TerrainPresetSchema,
 } from "./lib/ground";
 import {
-  formatMetres,
   formatOhms,
   formatScalar,
   formatSwr,
-  reflectionCoefficient,
 } from "./lib/format";
 import { richardsonExtrap } from "./lib/math";
 import {
@@ -65,21 +63,16 @@ import {
   type SchemaParamSpec,
 } from "./lib/params";
 import {
-  cross3,
-  dot3,
   MOBILE_SCREENS,
   PROJECTIONS,
   VIEWS,
   type Projection,
-  type Vec3,
   type View,
 } from "./lib/view";
 import type {
   ConvergeData,
-  FeedEntry,
   MeasuredData,
   NormCheckData,
-  PatternCuts,
   SolveRequest,
   SolveResponse,
   SweepData,
@@ -91,6 +84,19 @@ import { BandDropdown } from "./components/params/BandDropdown";
 import { GeometryCombobox } from "./components/params/GeometryCombobox";
 import { Knob } from "./components/params/Knob";
 import { ParamForm } from "./components/params/ParamForm";
+import {
+  cutsWsSend,
+  flushCutsWsPending,
+  resolveCutsWsMessage,
+  setCutsWsSend,
+  type CutsWsMessage,
+} from "./components/charts/cuts";
+import { CurrentCanvas } from "./components/charts/CurrentCanvas";
+import { FarFieldChart } from "./components/charts/FarFieldChart";
+import { GHOST_COLOR_COUNT, GHOST_FALLBACK_RGB } from "./components/charts/palette";
+import { SmithChart } from "./components/charts/SmithChart";
+import type { PatternMetrics, PinnedPattern } from "./components/charts/types";
+import { ThemeContext, type Theme } from "./components/hooks";
 
 // One advisory finding from the design screener (what a design does that a
 // typical one doesn't), attached to a trust-required entry.
@@ -180,7 +186,7 @@ function ResultPanel({
 // are directly comparable when the user switches slots.
 const CONVERGE_N_VALUES: number[] = [8, 12, 17, 24, 34, 48, 68];
 
-type PatternData = {
+export type PatternData = {
   theta_deg: number[];
   phi_deg: number[];
   gain_dbi: number[][];
@@ -311,8 +317,6 @@ function useThumbColumnSize(
   return size;
 }
 
-type Theme = "light" | "dark";
-const ThemeContext = createContext<Theme>("light");
 // Theme is global (owned by the shell) but the toggle button lives in each
 // session's sidebar header; sessions reach the setter through this context so
 // the single button drives the one shared theme.
@@ -2708,12 +2712,12 @@ function DesignSession({ id, active }: { id: number; active: boolean }) {
       return true;
     };
     const dropCutsSender = () => {
-      if (cutsWsSend === cutsSender) cutsWsSend = null;
+      if (cutsWsSend === cutsSender) setCutsWsSend(null);
       flushCutsWsPending();
     };
     ws.onopen = () => {
       setStatus("open");
-      cutsWsSend = cutsSender;
+      setCutsWsSend(cutsSender);
       // A prior socket's in-flight responses can never arrive on this new one.
       // Treat everything sent so far as received so `solving` can't stick true,
       // drop stale RTT timers, then send fresh current state. StrictMode and HMR
@@ -4710,19 +4714,6 @@ function ViewPanel({
   );
 }
 
-type FarFieldCut = "xy" | "yz";
-
-// Scalar far-field metrics from /pattern_metrics, shown in the compare table.
-type PatternMetrics = {
-  peak_gain_dbi: number;
-  takeoff_deg: number;
-  azimuth_deg: number;
-  front_to_back_db: number;
-  az_beamwidth_deg: number;
-  el_beamwidth_deg: number;
-  measurement_freq_mhz?: number;
-};
-
 // Fetch the scalar far-field metrics for a request (peak gain, takeoff, F/B,
 // beamwidths). Returns null when the design can't be evaluated or on error.
 async function fetchMetrics(req: SolveRequest): Promise<PatternMetrics | null> {
@@ -4738,27 +4729,6 @@ async function fetchMetrics(req: SolveRequest): Promise<PatternMetrics | null> {
     return null;
   }
 }
-
-// A pinned far-field snapshot: the full solve response (so its cut traces
-// recompute through the same math as the live one, in whatever cut the user is
-// viewing) plus a label, and the metrics fetched for the table. Pins live in
-// the shell and are shared across sessions (see PinsContext), so they survive
-// design switches and tab closes — you can overlay one antenna's pattern on
-// another's, including a design open in a different tab.
-type PinnedPattern = {
-  id: string;
-  label: string;
-  result: SolveResponse;
-  metrics: PatternMetrics | null;
-  // Whether the ghost overlay is drawn. A disabled pin keeps its table row
-  // (dimmed, metrics still readable) — that's the point of disable vs delete.
-  enabled: boolean;
-  // Fixed GHOST_COLORS slot, assigned at pin time. Stored — not the array
-  // index — because the chart draws a filtered (enabled-only) list while the
-  // table draws all pins; positional colors would desynchronize the two, and
-  // already shifted every later pin's color on delete.
-  colorIdx: number;
-};
 
 function PatternCompareTable({
   live,
@@ -4869,1936 +4839,4 @@ function PatternCompareTable({
       </tbody>
     </table>
   );
-}
-
-// --- Server-side polar cuts (issue #547) -----------------------------------
-// The per-direction cut physics lives in server.py (_pattern_cuts); every
-// solve response arrives with `cuts` attached at the request's angles, and
-// new angles come from the stateless POST /cuts endpoint. The pieces below
-// cache those fetches so the azimuth + elevation charts (and thumbnails)
-// sharing a solve don't duplicate round trips.
-
-// Identity for solve-response objects (immutable snapshots), used to key the
-// cuts cache. WeakMap so ids die with their solves.
-let cutsIdSeq = 0;
-const cutsIds = new WeakMap<SolveResponse, number>();
-// Resolved /cuts responses by `${solveId}:${azElev}:${elevAz}`. Insertion-
-// ordered and capped — a long session of slider drags would otherwise grow
-// it without bound.
-const cutsCache = new Map<string, PatternCuts>();
-const CUTS_CACHE_MAX = 256;
-// In-flight fetch dedup — both chart instances ask for the same key at once.
-const cutsInFlight = new Map<string, Promise<PatternCuts | null>>();
-// Freshest trace known per solve, at whatever angles — drawn while the fetch
-// for the current angles is in flight so drags never blank the chart.
-const cutsLatest = new WeakMap<SolveResponse, PatternCuts>();
-
-function cutsKey(
-  result: SolveResponse,
-  azElevDeg: number,
-  elevAzDeg: number,
-): string {
-  let id = cutsIds.get(result);
-  if (id === undefined) {
-    id = ++cutsIdSeq;
-    cutsIds.set(result, id);
-  }
-  return `${id}:${azElevDeg}:${elevAzDeg}`;
-}
-
-// The cuts for a solve at exactly these angles, or null if not yet known.
-function cachedCuts(
-  result: SolveResponse,
-  azElevDeg: number,
-  elevAzDeg: number,
-): PatternCuts | null {
-  const attached = result.cuts;
-  if (
-    attached &&
-    attached.az_elev_deg === azElevDeg &&
-    attached.elev_az_deg === elevAzDeg
-  ) {
-    return attached;
-  }
-  return cutsCache.get(cutsKey(result, azElevDeg, elevAzDeg)) ?? null;
-}
-
-// Cuts over the live /ws socket (issue #551): when the socket is open it
-// doubles as the cuts transport — a ~100-byte {_kind:"cuts", solve_id}
-// message replaces the full-body POST (warm ws measured ~2× cheaper than
-// HTTP through a tunnel, and the server answers from its cuts-source
-// cache). The socket effect registers a sender here on open and clears it
-// on close; fetchCuts falls back to HTTP when it's absent or doesn't
-// answer.
-let cutsWsSend: ((msg: string) => boolean) | null = null;
-// Outcome of a ws cuts round trip. "miss" (server says unknown id) skips
-// the pointless HTTP id retry — the same cache would 404 — and goes
-// straight to the full-body backstop; "unavailable" (no socket, timeout,
-// socket died) still tries the tiny HTTP id request first.
-type CutsWsReply =
-  | { status: "ok"; cuts: PatternCuts }
-  | { status: "miss" }
-  | { status: "unavailable" };
-const cutsWsPending = new Map<string, (reply: CutsWsReply) => void>();
-// Generous vs the ~100 ms worst-case big-mesh cut compute: a timeout only
-// fires when the socket is wedged, and the HTTP fallback then still
-// delivers — slower, never wrong.
-const CUTS_WS_TIMEOUT_MS = 1500;
-
-function cutsWsPendingKey(solveId: string, az: number, el: number): string {
-  return `${solveId}:${az}:${el}`;
-}
-
-function requestCutsViaWs(
-  solveId: string,
-  azElevDeg: number,
-  elevAzDeg: number,
-): Promise<CutsWsReply> {
-  const send = cutsWsSend;
-  if (!send) return Promise.resolve({ status: "unavailable" });
-  const key = cutsWsPendingKey(solveId, azElevDeg, elevAzDeg);
-  return new Promise((resolve) => {
-    const timer = window.setTimeout(() => {
-      cutsWsPending.delete(key);
-      resolve({ status: "unavailable" });
-    }, CUTS_WS_TIMEOUT_MS);
-    cutsWsPending.set(key, (reply) => {
-      window.clearTimeout(timer);
-      cutsWsPending.delete(key);
-      resolve(reply);
-    });
-    if (
-      !send(
-        JSON.stringify({
-          _kind: "cuts",
-          solve_id: solveId,
-          az_elev_deg: azElevDeg,
-          elev_az_deg: elevAzDeg,
-        }),
-      )
-    ) {
-      window.clearTimeout(timer);
-      cutsWsPending.delete(key);
-      resolve({ status: "unavailable" });
-    }
-  });
-}
-
-/** Server cuts message arriving on the /ws socket — routed here by the
- *  socket effect's onmessage before any solve handling. */
-type CutsWsMessage = {
-  _kind: "cuts";
-  solve_id: string;
-  az_elev_deg: number;
-  elev_az_deg: number;
-  ok: boolean;
-  cuts?: PatternCuts;
-};
-
-function resolveCutsWsMessage(data: CutsWsMessage): void {
-  const key = cutsWsPendingKey(data.solve_id, data.az_elev_deg, data.elev_az_deg);
-  const pending = cutsWsPending.get(key);
-  if (!pending) return; // timed out / superseded — fallback already running
-  pending(data.ok && data.cuts ? { status: "ok", cuts: data.cuts } : { status: "miss" });
-}
-
-/** Socket died: nothing pending will ever be answered on it. Resolving as
- *  "unavailable" sends every waiter down the HTTP path immediately instead
- *  of eating the full timeout. (A rare pending riding a *newer* socket gets
- *  flushed too — it just falls back to HTTP: slower, never wrong.) */
-function flushCutsWsPending(): void {
-  for (const pending of Array.from(cutsWsPending.values())) {
-    pending({ status: "unavailable" });
-  }
-}
-
-function fetchCuts(
-  result: SolveResponse,
-  azElevDeg: number,
-  elevAzDeg: number,
-): Promise<PatternCuts | null> {
-  const key = cutsKey(result, azElevDeg, elevAzDeg);
-  const inFlight = cutsInFlight.get(key);
-  if (inFlight) return inFlight;
-  const postCuts = (body: object): Promise<Response> =>
-    fetch("/cuts", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        ...body,
-        az_elev_deg: azElevDeg,
-        elev_az_deg: elevAzDeg,
-      }),
-    });
-  const p = (async (): Promise<PatternCuts | null> => {
-    // Transport ladder (issue #551): ws id → HTTP id → HTTP full body.
-    // Every rung is strictly a fallback of the one above; the full-body
-    // POST remains the correctness backstop (it's how pre-#551 responses
-    // and pins from dead server sessions resolve).
-    const solveId = result.solve_id;
-    if (solveId) {
-      const viaWs = await requestCutsViaWs(solveId, azElevDeg, elevAzDeg);
-      if (viaWs.status === "ok") return viaWs.cuts;
-      if (viaWs.status === "unavailable") {
-        const r = await postCuts({ solve_id: solveId });
-        if (r.ok) return (await r.json()) as PatternCuts;
-        if (r.status !== 404) return null; // 400: cuts genuinely unsupported
-      }
-      // ws said "miss" (or HTTP id 404'd): the server lost this solve —
-      // only the full body can answer now.
-    }
-    const r = await postCuts({ solve: result });
-    return r.ok ? ((await r.json()) as PatternCuts) : null;
-  })()
-    .then((cuts) => {
-      if (cuts) {
-        if (cutsCache.size >= CUTS_CACHE_MAX) {
-          // Evict the oldest half (Maps iterate in insertion order).
-          let drop = CUTS_CACHE_MAX >> 1;
-          for (const k of Array.from(cutsCache.keys())) {
-            if (drop-- <= 0) break;
-            cutsCache.delete(k);
-          }
-        }
-        cutsCache.set(key, cuts);
-      }
-      return cuts;
-    })
-    .catch(() => null)
-    .finally(() => cutsInFlight.delete(key));
-  cutsInFlight.set(key, p);
-  return p;
-}
-
-// Resolve the cut traces for a list of solves (live + pinned ghosts) at the
-// current angles. Synchronous when a solve already carries or has cached the
-// right angles; otherwise the freshest known trace is returned immediately
-// (stale-while-refetch) and a debounced POST /cuts brings the real one — the
-// debounce eats the flood of intermediate angles a slider drag produces.
-function useCutTraces(
-  results: readonly (SolveResponse | null)[],
-  azElevDeg: number,
-  elevAzDeg: number,
-): (PatternCuts | null)[] {
-  const [, setFetchTick] = useState(0); // re-render when a fetch lands
-  const wantKey = results
-    .map((r) => (r ? cutsKey(r, azElevDeg, elevAzDeg) : "-"))
-    .join("|");
-  useEffect(() => {
-    const missing = results.filter(
-      (r): r is SolveResponse => !!r && !cachedCuts(r, azElevDeg, elevAzDeg),
-    );
-    if (missing.length === 0) return;
-    let cancelled = false;
-    // When every missing trace can go over the ws id path (~100 bytes, and
-    // the server squashes per-solve latest-wins), a much tighter debounce
-    // makes cut drags feel live; the original 120 ms guard stays for the
-    // full-body HTTP fallback, whose requests are 10 KB–300 KB (issue #551).
-    const delay =
-      cutsWsSend && missing.every((r) => r.solve_id) ? 30 : 120;
-    const h = window.setTimeout(() => {
-      Promise.all(missing.map((r) => fetchCuts(r, azElevDeg, elevAzDeg))).then(
-        () => {
-          if (!cancelled) setFetchTick((t) => t + 1);
-        },
-      );
-    }, delay);
-    return () => {
-      cancelled = true;
-      window.clearTimeout(h);
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [wantKey]);
-  return results.map((r) => {
-    if (!r) return null;
-    const exact = cachedCuts(r, azElevDeg, elevAzDeg);
-    if (exact) {
-      cutsLatest.set(r, exact);
-      return exact;
-    }
-    return cutsLatest.get(r) ?? r.cuts ?? null;
-  });
-}
-
-// One chart's trace from a cuts payload: the dBi samples for the requested
-// cut plus their peak (for the adaptive radial scale and the annotation).
-// Null when there's nothing to draw (no cuts, or everything at the floor).
-function traceFor(
-  cuts: PatternCuts | null,
-  cut: FarFieldCut,
-): { dbi: number[]; peakDbi: number } | null {
-  if (!cuts) return null;
-  const dbi = cut === "xy" ? cuts.azimuth : cuts.elevation;
-  let peak = -Infinity;
-  for (const d of dbi) if (d > peak) peak = d;
-  if (!(peak > cuts.floor_dbi)) return null;
-  return { dbi, peakDbi: peak };
-}
-
-function FarFieldChart({
-  result,
-  pattern,
-  pinned,
-  size,
-  cut,
-  azElevDeg,
-  elevAzDeg,
-  fineNorm,
-}: {
-  result: SolveResponse | null;
-  pattern: PatternData | null;
-  pinned: PinnedPattern[];
-  size: number;
-  cut: FarFieldCut;
-  azElevDeg: number;
-  elevAzDeg: number;
-  /** Field-side gain norm from the dwell-triggered norm check (the pattern
-   *  renormalised by its own integrated radiated power instead of the input
-   *  power the live norm uses). When set, that pattern is overlaid dotted —
-   *  the norm is a scalar multiplier, so it is the live trace shifted
-   *  radially by 10·log10(fineNorm/liveNorm). Overlap ⇒ the solve conserves
-   *  power; a visible gap ⇒ the solver's discretisation error. */
-  fineNorm?: number | null;
-}) {
-  const theme = useContext(ThemeContext); // repaint on theme toggle (dep below)
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-
-  // Server-computed cut traces for the live solve + enabled pins (issue
-  // #547): synchronous when the angles match what each solve shipped with,
-  // POST /cuts (debounced, cached) after the user drags a cut slider.
-  // Disabled pins draw no ghost and don't stretch the radial scale.
-  const enabledPins = pinned.filter((p) => p.enabled);
-  const cutTraces = useCutTraces(
-    [result, ...enabledPins.map((p) => p.result)],
-    azElevDeg,
-    elevAzDeg,
-  );
-  // Draw-effect dep: changes when a fetched trace replaces a stale one (the
-  // solve identities and angles are already deps of their own).
-  const cutTracesKey = cutTraces
-    .map((t) => (t ? `${t.az_elev_deg},${t.elev_az_deg}` : "-"))
-    .join("|");
-
-  useEffect(() => {
-    const canvas = canvasRef.current;
-    if (!canvas) return;
-    const ctx = canvas.getContext("2d");
-    if (!ctx) return;
-
-    const dpr = window.devicePixelRatio || 1;
-    canvas.width = Math.floor(size * dpr);
-    canvas.height = Math.floor(size * dpr);
-    canvas.style.width = `${size}px`;
-    canvas.style.height = `${size}px`;
-    ctx.setTransform(dpr, 0, 0, dpr, 0, 0);
-
-    const PC = plotColors();
-
-    ctx.fillStyle = PC.bg;
-    ctx.fillRect(0, 0, size, size);
-
-    const cx = size / 2;
-    const cy = size / 2;
-    const R = size / 2 - 14;
-
-    // Azimuth cut: cone above horizon at elevation azElevDeg. With ground
-    // off, the conventional setting is 0° (the xy plane). With ground on,
-    // 0° is grazing and Fresnel kills the pattern, so something like 15°
-    // gives a useful view — the slider lets the user pick.
-    const azElevRad = (azElevDeg * Math.PI) / 180;
-    const azSinT = Math.cos(azElevRad); // sin(polar θ from +z) = cos(elevation)
-    const azCosT = Math.sin(azElevRad); // cos(polar θ) = sin(elevation)
-    // Elevation cut: vertical great circle through azimuth bearing elevAzDeg.
-    // t=0 lies at +elevAz horizon; t=π/2 is zenith; t=π is at the opposite
-    // horizon; t=3π/2 is nadir (below ground, zeroed when ground is on).
-    const elevAzRad = (elevAzDeg * Math.PI) / 180;
-    const elevAzCos = Math.cos(elevAzRad);
-    const elevAzSin = Math.sin(elevAzRad);
-
-    // Slice every trace (live + any pinned ghosts) up front, so the radial
-    // scale below can expand to fit the highest-gain lobe on screen.
-    const liveTrace = traceFor(cutTraces[0], cut);
-    const ghosts = enabledPins.map((p, i) => ({
-      colorIdx: p.colorIdx,
-      trace: traceFor(cutTraces[i + 1], cut),
-    }));
-
-    // Radial axis: absolute directivity in dBi. Origin is a fixed −20 dBi
-    // floor. The outer edge is +10 dBi by default, but expands to fit the peak
-    // of the highest-gain trace (plus 1 dB headroom) so a high-gain array's
-    // lobe renders in full instead of drawing past the edge and clipping — the
-    // thumbnails escaped this only because their tiny radius left slack inside
-    // the margin. Labeled rings sit at +6/0/−6/−12/−18 (all inside any top).
-    const DBI_FLOOR = -20;
-    const peaks: number[] = [];
-    if (liveTrace) peaks.push(liveTrace.peakDbi);
-    for (const gh of ghosts) if (gh.trace) peaks.push(gh.trace.peakDbi);
-    // Norm-check overlay: the norm scales the whole pattern, so switching to
-    // the field-side norm shifts every dBi by this constant. null when the
-    // check is off or the live result carries no norm to compare against.
-    const liveNorm = result?.directivity_norm;
-    const gridDeltaDb =
-      fineNorm && fineNorm > 0 && liveNorm && liveNorm > 0
-        ? 10 * Math.log10(fineNorm / liveNorm)
-        : null;
-    // Let the radial scale grow to fit the shifted overlay when it lands higher.
-    if (liveTrace && gridDeltaDb != null) peaks.push(liveTrace.peakDbi + gridDeltaDb);
-    const maxPeak = peaks.filter(Number.isFinite).reduce((a, b) => Math.max(a, b), 10);
-    const DBI_TOP = Math.max(10, Math.ceil(maxPeak + 1));
-    const DB_SPAN = DBI_TOP - DBI_FLOOR;
-    // Clamp to [0, 1]: a lobe at/above the top sits on the rim instead of
-    // drawing past R and clipping against the canvas edge.
-    const dbiToFrac = (db: number) =>
-      Math.max(0, Math.min(1, (db - DBI_FLOOR) / DB_SPAN));
-    ctx.strokeStyle = PC.grid;
-    ctx.lineWidth = 0.6;
-    ctx.fillStyle = PC.labelDim;
-    ctx.font = "9px ui-monospace, monospace";
-    for (const db of [6, 0, -6, -12, -18]) {
-      const f = dbiToFrac(db);
-      ctx.beginPath();
-      ctx.arc(cx, cy, R * f, 0, 2 * Math.PI);
-      ctx.stroke();
-      ctx.fillText(`${db > 0 ? "+" : ""}${db}`, cx + 2, cy - R * f - 1);
-    }
-    ctx.beginPath();
-    ctx.moveTo(cx - R, cy);
-    ctx.lineTo(cx + R, cy);
-    ctx.moveTo(cx, cy - R);
-    ctx.lineTo(cx, cy + R);
-    ctx.stroke();
-
-    // Axis labels: xy cut uses world x/y around the rim; yz cut shows the
-    // azimuth bearing on the horizontal pair and zenith/nadir on vertical.
-    ctx.fillStyle = PC.labelDim;
-    ctx.font = "10px ui-monospace, monospace";
-    const cutLabel =
-      cut === "xy"
-        ? `az @ ${azElevDeg}° elev (dBi)`
-        : `elev @ ${elevAzDeg}° az (dBi)`;
-    ctx.fillText(cutLabel, 6, 14);
-    ctx.fillStyle = PC.label;
-    if (cut === "xy") {
-      ctx.fillText("+x", cx + R - 14, cy + 11);
-      ctx.fillText("−x", cx - R + 2, cy + 11);
-      ctx.fillText("+y", cx - 8, cy - R + 12);
-      ctx.fillText("−y", cx - 7, cy + R - 2);
-    } else {
-      ctx.fillText("zen", cx - 9, cy - R + 12);
-      ctx.fillText("nad", cx - 9, cy + R - 2);
-    }
-
-    // Cross-reference: a single dashed spoke showing where the *other* cut
-    // slices this plot. The opposite side is implied by symmetry.
-    const markerStyle = PC.spoke;
-    {
-      const canvasAngleRad =
-        cut === "xy"
-          ? (elevAzDeg * Math.PI) / 180  // azimuth plot: elevation cut's bearing
-          : (azElevDeg * Math.PI) / 180; // elevation plot: azimuth cut's elevation
-      const cosA = Math.cos(canvasAngleRad);
-      const sinA = Math.sin(canvasAngleRad);
-      ctx.beginPath();
-      ctx.moveTo(cx, cy);
-      ctx.lineTo(cx + cosA * R, cy - sinA * R);
-      ctx.strokeStyle = markerStyle;
-      ctx.lineWidth = 0.8;
-      ctx.setLineDash([3, 3]);
-      ctx.stroke();
-      ctx.setLineDash([]);
-    }
-
-    if (!result) return;
-
-    // Terrain orientation marker: the terrain's characteristic bearing
-    // (water / downhill / cliff side), so orientation reads off the chart
-    // instead of being inferred from lobes — which can legitimately peak
-    // toward the OTHER side (a hillside's mid-angle lobes point uphill;
-    // downhill only wins below the first-lobe band).
-    const terrainMarker = result.ground_terrain?.marker;
-    if (terrainMarker) {
-      ctx.font = "10px ui-monospace, monospace";
-      if (cut === "xy") {
-        // Inward tick + label at the bearing; dimmer label opposite.
-        const a = (terrainMarker.bearing_deg * Math.PI) / 180;
-        const ca = Math.cos(a);
-        const sa = Math.sin(a);
-        ctx.strokeStyle = PC.labelStrong;
-        ctx.lineWidth = 1.5;
-        ctx.beginPath();
-        ctx.moveTo(cx + ca * R, cy - sa * R);
-        ctx.lineTo(cx + ca * (R - 8), cy - sa * (R - 8));
-        ctx.stroke();
-        const place = (label: string, dirx: number, diry: number) => {
-          const tw = ctx.measureText(label).width;
-          // Sit the label just inside the rim along the bearing, nudged
-          // toward the centre so it doesn't collide with the axis labels.
-          ctx.fillText(
-            label,
-            cx + dirx * (R - 24) - tw / 2,
-            cy - diry * (R - 24) + 4,
-          );
-        };
-        ctx.fillStyle = PC.labelStrong;
-        place(terrainMarker.label, ca, sa);
-        ctx.fillStyle = PC.labelDim;
-        place(terrainMarker.opposite, -ca, -sa);
-      } else {
-        // Elevation cut: label which terrain side each horizon points into
-        // (the right rim is the cut bearing elevAzDeg).
-        const rel =
-          ((((elevAzDeg - terrainMarker.bearing_deg) % 360) + 540) % 360) -
-          180;
-        const rightLabel =
-          Math.abs(rel) <= 90 ? terrainMarker.label : terrainMarker.opposite;
-        const leftLabel =
-          Math.abs(rel) <= 90 ? terrainMarker.opposite : terrainMarker.label;
-        ctx.fillStyle = PC.labelStrong;
-        ctx.fillText(
-          rightLabel,
-          cx + R - ctx.measureText(rightLabel).width - 2,
-          cy - 5,
-        );
-        ctx.fillStyle = PC.labelDim;
-        ctx.fillText(leftLabel, cx - R + 2, cy - 5);
-      }
-    }
-
-    // Draw one dBi trace around the polar cut (sample i at t = 2π·i/n, the
-    // server cuts' parameterisation). The live lobe closes + fills; pinned
-    // ghosts are an open dashed stroke so the live trace reads on top.
-    const strokeTrace = (
-      dbi: number[],
-      o: { stroke: string; fill?: string; width: number; dash?: number[] },
-    ) => {
-      const n = dbi.length;
-      ctx.beginPath();
-      for (let pi = 0; pi <= n; pi++) {
-        const t = (2 * Math.PI * pi) / n;
-        const frac = dbiToFrac(dbi[pi % n]);
-        const px = cx + Math.cos(t) * frac * R;
-        // Canvas y flips: +y on canvas is down, so we negate to put +y at top.
-        const py = cy - Math.sin(t) * frac * R;
-        if (pi === 0) ctx.moveTo(px, py);
-        else ctx.lineTo(px, py);
-      }
-      ctx.closePath();
-      if (o.fill) {
-        ctx.fillStyle = o.fill;
-        ctx.fill();
-      }
-      if (o.dash) ctx.setLineDash(o.dash);
-      ctx.strokeStyle = o.stroke;
-      ctx.lineWidth = o.width;
-      ctx.stroke();
-      if (o.dash) ctx.setLineDash([]);
-    };
-
-    // Pinned ghosts first (dimmed, dashed), so the live lobe sits on top. Each
-    // shares the adaptive radial scale computed above, so it tracks the cut and
-    // angle sliders just like the live trace.
-    for (const gh of ghosts) {
-      if (!gh.trace) continue;
-      strokeTrace(gh.trace.dbi, {
-        stroke: `rgba(${ghostRgb(gh.colorIdx)}, 0.8)`,
-        width: 1,
-        dash: [5, 3],
-      });
-    }
-
-    // Live lobe (filled).
-    if (!liveTrace) return;
-    strokeTrace(liveTrace.dbi, {
-      stroke: `rgba(${PC.lobeRgb}, 0.9)`,
-      fill: `rgba(${PC.lobeRgb}, 0.12)`,
-      width: 1.5,
-    });
-
-    // Fine-grid norm overlay (dotted, same lobe hue): the live trace shifted
-    // radially by the constant dB offset. Sits exactly on the solid lobe when
-    // the adaptive grid was fine enough; a visible gap is the grid error. Drawn
-    // open (no fill) so the solid lobe still reads underneath.
-    if (gridDeltaDb != null) {
-      strokeTrace(
-        liveTrace.dbi.map((d) => d + gridDeltaDb),
-        { stroke: `rgba(${PC.lobeRgb}, 0.85)`, width: 1, dash: [2, 2] },
-      );
-    }
-
-    // NEC exact-pattern overlay (dashed cyan line) when available. Bilinear
-    // interpolation off the (θ, φ) grid; rays below horizon are skipped so
-    // the line breaks at the ground rather than wrapping to the origin.
-    if (pattern) {
-      const N_DIR = 180; // drawing resolution for the interpolated overlay
-      const nt = pattern.theta_deg.length;
-      const np_ = pattern.phi_deg.length;
-      const dTheta = pattern.theta_deg[1] - pattern.theta_deg[0];
-      const dPhi = pattern.phi_deg[1] - pattern.phi_deg[0];
-      const clip = (g: number) => (g < -100 ? -100 : g);
-
-      ctx.beginPath();
-      let started = false;
-      for (let pi = 0; pi <= N_DIR; pi++) {
-        const t = (2 * Math.PI * pi) / N_DIR;
-        const ct = Math.cos(t);
-        const st = Math.sin(t);
-        const rx = cut === "xy" ? azSinT * ct : elevAzCos * ct;
-        const ry = cut === "xy" ? azSinT * st : elevAzSin * ct;
-        const rz = cut === "xy" ? azCosT : st;
-        if (rz < -1e-9) { started = false; continue; }
-
-        const thetaDeg = (Math.acos(Math.max(-1, Math.min(1, rz))) * 180) / Math.PI;
-        let phiRad = Math.atan2(ry, rx);
-        if (phiRad < 0) phiRad += 2 * Math.PI;
-        const phiDeg = (phiRad * 180) / Math.PI;
-
-        const tf = Math.max(0, Math.min(nt - 1, thetaDeg / dTheta));
-        const pf = Math.max(0, Math.min(np_ - 1, phiDeg / dPhi));
-        const t0 = Math.floor(tf), t1 = Math.min(nt - 1, t0 + 1);
-        const p0 = Math.floor(pf), p1 = Math.min(np_ - 1, p0 + 1);
-        const ft = tf - t0, fp = pf - p0;
-        const g00 = clip(pattern.gain_dbi[t0][p0]);
-        const g01 = clip(pattern.gain_dbi[t0][p1]);
-        const g10 = clip(pattern.gain_dbi[t1][p0]);
-        const g11 = clip(pattern.gain_dbi[t1][p1]);
-        const dBi =
-          g00 * (1 - ft) * (1 - fp) +
-          g01 * (1 - ft) * fp +
-          g10 * ft * (1 - fp) +
-          g11 * ft * fp;
-
-        const frac = dbiToFrac(dBi);
-        const px = cx + Math.cos(t) * frac * R;
-        const py = cy - Math.sin(t) * frac * R;
-        if (!started) { ctx.moveTo(px, py); started = true; }
-        else ctx.lineTo(px, py);
-      }
-      ctx.strokeStyle = `rgba(${PC.necRgb}, 0.85)`;
-      ctx.lineWidth = 1;
-      ctx.setLineDash([4, 3]);
-      ctx.stroke();
-      ctx.setLineDash([]);
-
-      // Legend swatch + label, bottom-right.
-      ctx.fillStyle = `rgba(${PC.necRgb}, 0.9)`;
-      ctx.font = "10px ui-monospace, monospace";
-      const necText = "NEC rp_card";
-      const necTw = ctx.measureText(necText).width;
-      ctx.fillText(necText, size - necTw - 6, size - 6);
-    }
-
-    // Peak dBi annotation (top-right corner).
-    const peakDbi = liveTrace.peakDbi;
-    ctx.fillStyle = PC.labelStrong;
-    ctx.font = "10px ui-monospace, monospace";
-    const peakText = `peak ${peakDbi >= 0 ? "+" : ""}${peakDbi.toFixed(1)} dBi`;
-    const tw = ctx.measureText(peakText).width;
-    ctx.fillText(peakText, size - tw - 6, 14);
-    // cutTracesKey stands in for the fetched trace contents (see above); the
-    // other deps cover everything the draw reads directly.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [result, pattern, pinned, size, cut, azElevDeg, elevAzDeg, fineNorm, theme, cutTracesKey]);
-
-  return <canvas ref={canvasRef} className="farfield" />;
-}
-
-// Per-feed colors for multi-line Smith chart overlays. Feed 0 keeps the
-// existing single-feed blue so single-feed geometries are visually
-// unchanged; subsequent feeds use distinct hues that read well on the
-// dark background. Indices beyond this list wrap, but that's only
-// reachable on >4-feed geometries (none exist yet).
-const FEED_COLORS: [number, number, number][] = [
-  [118, 208, 255],  // blue (primary)
-  [255, 196, 102],  // amber
-  [140, 230, 140],  // green
-  [255, 130, 200],  // pink
-];
-
-// Swatch colors for pinned far-field ghost overlays, themed via CSS vars —
-// the dark theme's pastels are darker inks in light mode, where a 1px dashed
-// pastel stroke vanishes on the white canvas. Distinct from the live lobe
-// (orange) and the NEC overlay (cyan); they wrap past the 4th pin.
-const GHOST_COLOR_COUNT = 4;
-// Fallbacks match the dark-theme values in styles.css.
-const GHOST_FALLBACK_RGB = [
-  "140, 230, 140", // green
-  "255, 130, 200", // pink
-  "180, 160, 255", // violet
-  "120, 220, 220", // teal
-];
-// "r, g, b" for a pin's color slot in the current theme, for canvas strokes.
-// (The compare table instead inlines the CSS var so it rethemes live.)
-function ghostRgb(colorIdx: number): string {
-  const i = colorIdx % GHOST_COLOR_COUNT;
-  const v = getComputedStyle(document.documentElement)
-    .getPropertyValue(`--plot-ghost-${i}-rgb`)
-    .trim();
-  return v || GHOST_FALLBACK_RGB[i];
-}
-
-function feedColor(i: number, alpha = 0.85): string {
-  const [r, g, b] = FEED_COLORS[i % FEED_COLORS.length];
-  return `rgba(${r}, ${g}, ${b}, ${alpha})`;
-}
-
-// Sweep trail uses a darkened variant of each feed's primary color so the
-// current-Z marker reads as "you are here" against a dimmer "trail". With
-// two feeds at identical Z (e.g. the in-phase symmetric case) the two
-// primary markers stack on top of each other but stay distinguishable from
-// the sweep cloud underneath — without this they were indistinguishable.
-function feedSweepColor(i: number, alpha = 0.85): string {
-  const [r, g, b] = FEED_COLORS[i % FEED_COLORS.length];
-  const f = 0.55; // darken factor — empirically readable on the #0d1015 bg
-  return `rgba(${Math.round(r * f)}, ${Math.round(g * f)}, ${Math.round(b * f)}, ${alpha})`;
-}
-
-function SmithChart({
-  r,
-  x,
-  z0,
-  size,
-  sweep,
-  converge,
-  measured,
-  measFreqMhz,
-  running,
-  convergeRunning,
-  feeds,
-  multiFeed,
-}: {
-  r: number;
-  x: number;
-  z0: number;
-  size: number;
-  sweep: SweepData | null;
-  converge: ConvergeData | null;
-  /** Uploaded VNA measurement drawn against the modeled locus (issue #595). */
-  measured: MeasuredData | null;
-  measFreqMhz: number;
-  running: boolean;
-  convergeRunning: boolean;
-  /** Multi-feed geometries pass the per-feed Z list from the latest
-   *  solve so the chart can also render N centre dots, one per port. */
-  feeds?: FeedEntry[];
-  /** From the example descriptor's `multi_feed` flag — drives the
-   *  per-feed summary rows. Decoupled from feeds[].length so the chart
-   *  reflects antenna type rather than guessing from response shape. */
-  multiFeed: boolean;
-}) {
-  const theme = useContext(ThemeContext); // repaint on theme toggle (dep below)
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-
-  useEffect(() => {
-    const canvas = canvasRef.current;
-    if (!canvas) return;
-    const ctx = canvas.getContext("2d");
-    if (!ctx) return;
-
-    const dpr = window.devicePixelRatio || 1;
-    canvas.width = Math.floor(size * dpr);
-    canvas.height = Math.floor(size * dpr);
-    canvas.style.width = `${size}px`;
-    canvas.style.height = `${size}px`;
-    ctx.setTransform(dpr, 0, 0, dpr, 0, 0);
-
-    const PC = plotColors();
-
-    const cx = size / 2;
-    const cy = size / 2;
-    const R = size / 2 - 10;
-
-    ctx.fillStyle = PC.bg;
-    ctx.fillRect(0, 0, size, size);
-
-    // Constant-r circles in the Γ plane.
-    // Each maps to a circle: center = (r/(r+1), 0), radius = 1/(r+1).
-    const rCircles: { r: number; label?: string }[] = [
-      { r: 0.2 },
-      { r: 0.5, label: "0.5" },
-      { r: 1, label: "1" },
-      { r: 2, label: "2" },
-      { r: 5 },
-    ];
-    ctx.strokeStyle = PC.grid;
-    ctx.lineWidth = 0.6;
-    for (const { r: rn } of rCircles) {
-      const cxN = rn / (rn + 1);
-      const radN = 1 / (rn + 1);
-      ctx.beginPath();
-      ctx.arc(cx + cxN * R, cy, radN * R, 0, 2 * Math.PI);
-      ctx.stroke();
-    }
-
-    // Constant-x arcs: center = (1, 1/x), radius = 1/|x|. Clip to unit disk.
-    const xArcs = [0.2, 0.5, 1, 2, 5];
-    ctx.save();
-    ctx.beginPath();
-    ctx.arc(cx, cy, R, 0, 2 * Math.PI);
-    ctx.clip();
-    for (const xn of xArcs) {
-      const arcCx = cx + R;
-      const rad = (1 / xn) * R;
-      // Inductive (X > 0)
-      ctx.beginPath();
-      ctx.arc(arcCx, cy - (1 / xn) * R, rad, 0, 2 * Math.PI);
-      ctx.stroke();
-      // Capacitive (X < 0)
-      ctx.beginPath();
-      ctx.arc(arcCx, cy + (1 / xn) * R, rad, 0, 2 * Math.PI);
-      ctx.stroke();
-    }
-    ctx.restore();
-
-    // Real axis
-    ctx.strokeStyle = PC.axis;
-    ctx.lineWidth = 0.8;
-    ctx.beginPath();
-    ctx.moveTo(cx - R, cy);
-    ctx.lineTo(cx + R, cy);
-    ctx.stroke();
-
-    // Outer boundary (|Γ| = 1)
-    ctx.strokeStyle = PC.axis;
-    ctx.lineWidth = 1.5;
-    ctx.beginPath();
-    ctx.arc(cx, cy, R, 0, 2 * Math.PI);
-    ctx.stroke();
-
-    // Z0 label at center
-    ctx.fillStyle = PC.labelDim;
-    ctx.font = "10px ui-monospace, monospace";
-    ctx.fillText(`Z₀ = ${z0}`, 6, 14);
-
-    // Reactance sign labels.
-    ctx.fillStyle = PC.labelDim;
-    ctx.fillText("+jX", cx + R - 24, cy - R + 14);
-    ctx.fillText("−jX", cx + R - 24, cy + R - 4);
-
-    // Sweep locus: one colored trajectory per feed (or just the primary
-    // for single-feed geometries). Multi-feed geometries (bowtie) ship
-    // per-feed Z arrays via sweep.feeds_z_re / feeds_z_im; when present
-    // we render one color-distinct trajectory per port instead of the
-    // single legacy blue locus. No connecting line — sparse samples
-    // make a piecewise polyline read as artificial kinks.
-    if (sweep && sweep.freqs_mhz.length > 1) {
-      const hasMulti =
-        !!sweep.feeds_z_re &&
-        !!sweep.feeds_z_im &&
-        sweep.feeds_z_re.length === sweep.freqs_mhz.length &&
-        sweep.feeds_z_re[0].length > 1;
-      const nFeeds = hasMulti ? sweep.feeds_z_re![0].length : 1;
-
-      // Z accessor per (feed index, sample index). Single-feed falls
-      // back to the top-level z_re/z_im (same as before this change).
-      const zAt = (fi: number, i: number) =>
-        hasMulti
-          ? { re: sweep.feeds_z_re![i][fi], im: sweep.feeds_z_im![i][fi] }
-          : { re: sweep.z_re[i], im: sweep.z_im[i] };
-
-      ctx.save();
-      ctx.beginPath();
-      ctx.arc(cx, cy, R, 0, 2 * Math.PI);
-      ctx.clip();
-      for (let fi = 0; fi < nFeeds; fi++) {
-        // Darkened color so the sweep cloud reads as a trail underneath
-        // the bright current-Z primary marker (drawn later, full color).
-        // Same convention for single- and multi-feed so the chart's
-        // visual grammar is uniform.
-        ctx.fillStyle = feedSweepColor(fi);
-        for (let i = 0; i < sweep.freqs_mhz.length; i++) {
-          const z = zAt(fi, i);
-          const g = reflectionCoefficient(z.re, z.im, z0);
-          const px = cx + g.gRe * R;
-          const py = cy - g.gIm * R;
-          ctx.beginPath();
-          ctx.arc(px, py, 1.5, 0, 2 * Math.PI);
-          ctx.fill();
-        }
-      }
-      ctx.restore();
-
-      // Endpoint markers per feed (low-freq filled, high-freq hollow) —
-      // also drawn in the darkened sweep color so they stay part of the
-      // trail and don't compete with the bright current-Z marker.
-      const drawEndpoint = (fi: number, idx: number, filled: boolean) => {
-        const z = zAt(fi, idx);
-        const g = reflectionCoefficient(z.re, z.im, z0);
-        const px = cx + g.gRe * R;
-        const py = cy - g.gIm * R;
-        const col = feedSweepColor(fi);
-        ctx.lineWidth = 1.2;
-        ctx.strokeStyle = col;
-        ctx.fillStyle = filled ? col : `rgba(${PC.bgRgb}, 0.95)`;
-        ctx.beginPath();
-        ctx.arc(px, py, 3, 0, 2 * Math.PI);
-        ctx.fill();
-        ctx.stroke();
-      };
-      for (let fi = 0; fi < nFeeds; fi++) {
-        drawEndpoint(fi, 0, true);
-        drawEndpoint(fi, sweep.freqs_mhz.length - 1, false);
-      }
-
-      // Freq range label across the bottom of the panel.
-      ctx.fillStyle = PC.labelBright;
-      ctx.font = "10px ui-monospace, monospace";
-      const fLoTxt = sweep.freqs_mhz[0].toFixed(2);
-      const fHiTxt = sweep.freqs_mhz[sweep.freqs_mhz.length - 1].toFixed(2);
-      const txt = `${fLoTxt} → ${fHiTxt} MHz`;
-      ctx.fillText(txt, size - 6 - ctx.measureText(txt).width, size - 6);
-
-    }
-
-    // Measured overlay (issue #595): the locus a VNA actually saw, against the
-    // one the model predicts. The measurement arrives as impedance and goes
-    // through the same reflectionCoefficient() as the solved Z, so the file's
-    // own calibration reference is absorbed by the conversion — a 75 Ω
-    // measurement lands correctly on a 50 Ω chart with no special case.
-    if (measured && measured.freqs_mhz.length > 0) {
-      // Clip to the swept band so both loci span the same frequencies (the CLI
-      // overlay applies the same rule). With no sweep on screen there is
-      // nothing to clip against, so the whole measurement is drawn.
-      const mf = measured.freqs_mhz;
-      const swept = sweep && sweep.freqs_mhz.length > 1;
-      const fLo = swept ? sweep!.freqs_mhz[0] : -Infinity;
-      const fHi = swept ? sweep!.freqs_mhz[sweep!.freqs_mhz.length - 1] : Infinity;
-      const idx: number[] = [];
-      for (let i = 0; i < mf.length; i++) {
-        if (mf[i] >= fLo && mf[i] <= fHi) idx.push(i);
-      }
-      const mGamma = (i: number) =>
-        reflectionCoefficient(measured.z_re[i], measured.z_im[i], z0);
-      ctx.font = "10px ui-monospace, monospace";
-      ctx.fillStyle = PC.measured;
-      if (idx.length === 0) {
-        // Disjoint bands. Say so — the user picked a file and would otherwise
-        // see the chart not change at all.
-        const txt = `measured ${mf[0].toFixed(2)}–${mf[mf.length - 1].toFixed(2)} MHz: outside the sweep`;
-        ctx.fillText(txt, size - 6 - ctx.measureText(txt).width, size - 20);
-      } else {
-        ctx.save();
-        ctx.beginPath();
-        ctx.arc(cx, cy, R, 0, 2 * Math.PI);
-        ctx.clip();
-        // Dashed, to read as "other source" next to the solid convergence
-        // trail and the scattered sweep dots.
-        ctx.setLineDash([4, 3]);
-        ctx.strokeStyle = PC.measured;
-        ctx.lineWidth = 1.4;
-        ctx.beginPath();
-        for (let k = 0; k < idx.length; k++) {
-          const g = mGamma(idx[k]);
-          const px = cx + g.gRe * R;
-          const py = cy - g.gIm * R;
-          if (k === 0) ctx.moveTo(px, py);
-          else ctx.lineTo(px, py);
-        }
-        ctx.stroke();
-        ctx.setLineDash([]);
-        ctx.restore();
-
-        // Endpoint markers matching the sweep trail's grammar: low-frequency
-        // end filled, high-frequency end hollow, so the locus has a direction.
-        for (const [k, filled] of [[0, true], [idx.length - 1, false]] as const) {
-          const g = mGamma(idx[k]);
-          ctx.lineWidth = 1.2;
-          ctx.strokeStyle = PC.measured;
-          ctx.fillStyle = filled ? PC.measured : `rgba(${PC.bgRgb}, 0.95)`;
-          ctx.beginPath();
-          ctx.arc(cx + g.gRe * R, cy - g.gIm * R, 3, 0, 2 * Math.PI);
-          ctx.fill();
-          ctx.stroke();
-        }
-
-        // Label above the sweep's freq-range line, with the span actually
-        // drawn — which is the measurement's own band when it is narrower than
-        // the sweep, and says as much when the file reaches past it.
-        ctx.fillStyle = PC.measured;
-        const dLo = mf[idx[0]].toFixed(2);
-        const dHi = mf[idx[idx.length - 1]].toFixed(2);
-        const partial = idx.length < mf.length ? " (clipped)" : "";
-        const txt = `measured: ${measured.label}  ${dLo} → ${dHi} MHz${partial}`;
-        ctx.fillText(txt, size - 6 - ctx.measureText(txt).width, size - 20);
-      }
-    }
-
-    if (running) {
-      ctx.fillStyle = PC.label;
-      ctx.font = "10px ui-monospace, monospace";
-      ctx.fillText("sweeping…", 6, size - 6);
-    }
-
-    // Convergence locus: Z(N) trajectory as N increases, drawn as a
-    // connected polyline per feed so the sequence direction reads as
-    // motion (vs. the freq sweep's unconnected scatter). Each feed gets
-    // its own bright color from the feed palette — the line shape
-    // distinguishes the convergence trail from the scattered freq-sweep
-    // dots, and the bright-vs-dim brightness distinguishes the bright
-    // current-Z marker from the trail's interior dots. Smallest-N point
-    // gets a hollow ring; largest-N gets a filled disc; Richardson-
-    // extrapolated Z* gets a diamond (primary feed only).
-    if (converge && converge.n_values.length >= 1) {
-      const cHasMulti =
-        !!converge.feeds_z_re &&
-        !!converge.feeds_z_im &&
-        converge.feeds_z_re.length === converge.n_values.length &&
-        converge.feeds_z_re[0].length > 1;
-      const cNFeeds = cHasMulti ? converge.feeds_z_re![0].length : 1;
-      const czAt = (fi: number, i: number) =>
-        cHasMulti
-          ? { re: converge.feeds_z_re![i][fi], im: converge.feeds_z_im![i][fi] }
-          : { re: converge.z_re[i], im: converge.z_im[i] };
-
-      ctx.save();
-      ctx.beginPath();
-      ctx.arc(cx, cy, R, 0, 2 * Math.PI);
-      ctx.clip();
-      for (let fi = 0; fi < cNFeeds; fi++) {
-        ctx.strokeStyle = feedColor(fi);
-        ctx.lineWidth = 1.2;
-        ctx.beginPath();
-        for (let i = 0; i < converge.n_values.length; i++) {
-          const z = czAt(fi, i);
-          const g = reflectionCoefficient(z.re, z.im, z0);
-          const px = cx + g.gRe * R;
-          const py = cy - g.gIm * R;
-          if (i === 0) ctx.moveTo(px, py);
-          else ctx.lineTo(px, py);
-        }
-        ctx.stroke();
-
-        // Per-N dots along the trajectory.
-        ctx.fillStyle = feedColor(fi);
-        for (let i = 0; i < converge.n_values.length; i++) {
-          const z = czAt(fi, i);
-          const g = reflectionCoefficient(z.re, z.im, z0);
-          const px = cx + g.gRe * R;
-          const py = cy - g.gIm * R;
-          ctx.beginPath();
-          ctx.arc(px, py, 1.8, 0, 2 * Math.PI);
-          ctx.fill();
-        }
-      }
-      ctx.restore();
-
-      // Endpoint markers per feed: smallest-N hollow, largest-N filled.
-      const drawNEndpoint = (fi: number, idx: number, filled: boolean) => {
-        const z = czAt(fi, idx);
-        const g = reflectionCoefficient(z.re, z.im, z0);
-        const px = cx + g.gRe * R;
-        const py = cy - g.gIm * R;
-        const col = feedColor(fi);
-        ctx.lineWidth = 1.2;
-        ctx.strokeStyle = col;
-        ctx.fillStyle = filled ? col : `rgba(${PC.bgRgb}, 0.95)`;
-        ctx.beginPath();
-        ctx.arc(px, py, 3, 0, 2 * Math.PI);
-        ctx.fill();
-        ctx.stroke();
-      };
-      for (let fi = 0; fi < cNFeeds; fi++) {
-        drawNEndpoint(fi, 0, false);
-        drawNEndpoint(fi, converge.n_values.length - 1, true);
-      }
-
-      // Richardson Z* markers — one diamond per feed, each in the
-      // matching bright feed color so the user can tell which trail
-      // extrapolates to which Z*. The diamond shape distinguishes the
-      // extrapolated value from the actual sampled per-N dots (small
-      // circles) and from the current-Z marker (larger outlined dot).
-      const drawExtrap = (
-        fi: number,
-        zRe: number | null,
-        zIm: number | null,
-      ) => {
-        if (zRe == null || zIm == null) return;
-        const ge = reflectionCoefficient(zRe, zIm, z0);
-        // Clip to the unit Smith disc — Richardson on a not-yet-converging
-        // series can fly outside |Γ|=1 in early frames.
-        const gMag = Math.hypot(ge.gRe, ge.gIm);
-        const k = gMag > 0.98 ? 0.98 / gMag : 1;
-        const px = cx + ge.gRe * R * k;
-        const py = cy - ge.gIm * R * k;
-        ctx.save();
-        ctx.translate(px, py);
-        ctx.rotate(Math.PI / 4);
-        ctx.fillStyle = feedColor(fi);
-        ctx.strokeStyle = feedColor(fi, 1.0);
-        ctx.lineWidth = 1.2;
-        ctx.beginPath();
-        ctx.rect(-4, -4, 8, 8);
-        ctx.fill();
-        ctx.stroke();
-        ctx.restore();
-      };
-      if (cHasMulti && converge.feeds_z_re_extrap && converge.feeds_z_im_extrap) {
-        for (let fi = 0; fi < cNFeeds; fi++) {
-          drawExtrap(
-            fi,
-            converge.feeds_z_re_extrap[fi],
-            converge.feeds_z_im_extrap[fi],
-          );
-        }
-      } else {
-        drawExtrap(0, converge.z_re_extrap, converge.z_im_extrap);
-      }
-
-    }
-    if (convergeRunning) {
-      ctx.fillStyle = PC.label;
-      ctx.font = "10px ui-monospace, monospace";
-      // Stack under the freq-sweep status if both are running.
-      const yOff = running ? 18 : 6;
-      ctx.fillText("converging…", 6, size - yOff);
-    }
-
-    // Current impedance marker(s). One bright dot per feed in the
-    // matching feed color, with a thin dark outline for visibility on
-    // top of the sweep cloud. Single- and multi-feed share this code
-    // path so the chart's visual grammar is uniform: dim color = trail
-    // (freq sweep), bright = "you are here." The previous golden +
-    // glow + line-from-centre treatment for single-feed is gone —
-    // single-feed and feed-0-of-multi-feed now look identical.
-    const markerPoints: Array<{ re: number; im: number; fi: number }> =
-      feeds && feeds.length > 0
-        ? feeds.map((f, fi) => ({ re: f.z_re, im: f.z_im, fi }))
-        : r > 0 || x !== 0
-          ? [{ re: r, im: x, fi: 0 }]
-          : [];
-    for (const m of markerPoints) {
-      if (m.re <= 0 && m.im === 0) continue;
-      const { gRe, gIm } = reflectionCoefficient(m.re, m.im, z0);
-      const px = cx + gRe * R;
-      const py = cy - gIm * R;
-      ctx.fillStyle = feedColor(m.fi);
-      ctx.beginPath();
-      ctx.arc(px, py, 4, 0, 2 * Math.PI);
-      ctx.fill();
-      ctx.strokeStyle = `rgba(${PC.bgRgb}, 0.85)`;
-      ctx.lineWidth = 1;
-      ctx.stroke();
-    }
-
-    // Top-left summary: one row per feed (when multi-feed) or one row
-    // total (single-feed). Each row gets:
-    //   [dim swatch][bright swatch]  feed N  Z* ≈ R + jX Ω
-    // Both swatches encode the per-feed color (dim = freq-sweep trail,
-    // bright = current-Z marker / convergence trail). Z* tacks on
-    // inline in the matching bright color when ≥3 converge samples
-    // have come in, so each trail has its own visibly-colored Z*
-    // readout next to its swatch — replacing the old single purple
-    // Z* line that only tracked feeds[0].
-    const summaryFeeds: Array<{
-      fi: number;
-      extrapRe: number | null;
-      extrapIm: number | null;
-    }> = [];
-    if (multiFeed && feeds && feeds.length > 1) {
-      for (let fi = 0; fi < feeds.length; fi++) {
-        const re = converge?.feeds_z_re_extrap?.[fi] ?? null;
-        const im = converge?.feeds_z_im_extrap?.[fi] ?? null;
-        summaryFeeds.push({ fi, extrapRe: re, extrapIm: im });
-      }
-    } else if (converge && converge.n_values.length >= 1) {
-      summaryFeeds.push({
-        fi: 0,
-        extrapRe: converge.z_re_extrap,
-        extrapIm: converge.z_im_extrap,
-      });
-    } else if (feeds && feeds.length === 1) {
-      // Sweep-only single-feed run: show the swatch row so the colors
-      // on the chart are explained even without a converge.
-      summaryFeeds.push({ fi: 0, extrapRe: null, extrapIm: null });
-    }
-    if (summaryFeeds.length > 0) {
-      ctx.font = "10px ui-monospace, monospace";
-      for (let row = 0; row < summaryFeeds.length; row++) {
-        const { fi, extrapRe, extrapIm } = summaryFeeds[row];
-        const ly = 12 + row * 14;
-        // Dim swatch (sweep trail color).
-        ctx.fillStyle = feedSweepColor(fi);
-        ctx.beginPath();
-        ctx.arc(12, ly, 3, 0, 2 * Math.PI);
-        ctx.fill();
-        // Bright swatch (current-Z / convergence-trail color).
-        ctx.fillStyle = feedColor(fi);
-        ctx.beginPath();
-        ctx.arc(20, ly, 3, 0, 2 * Math.PI);
-        ctx.fill();
-        // Feed label + inline Z* (when extrap available). Text color
-        // matches the bright feed color so the row's color identity
-        // ties back to the chart trails for that feed.
-        ctx.fillStyle = feedColor(fi);
-        let txt =
-          summaryFeeds.length > 1 ? `feed ${fi}` : "";
-        if (extrapRe != null && extrapIm != null) {
-          const sign = extrapIm >= 0 ? "+" : "−";
-          const zText = `Z* ≈ ${extrapRe.toFixed(2)} ${sign} j${Math.abs(extrapIm).toFixed(2)} Ω`;
-          txt = txt ? `${txt}  ${zText}` : zText;
-        }
-        if (txt) ctx.fillText(txt, 28, ly + 3);
-      }
-    }
-
-    // Bottom-left: N-range stays neutral since it's per-converge not
-    // per-feed. Sits above the converging / sweeping status indicators.
-    if (converge && converge.n_values.length >= 1) {
-      const nLo = converge.n_values[0];
-      const nHi = converge.n_values[converge.n_values.length - 1];
-      ctx.fillStyle = PC.labelBright;
-      ctx.font = "10px ui-monospace, monospace";
-      const baseY = running && convergeRunning ? size - 30
-        : running || convergeRunning ? size - 18
-        : size - 6;
-      ctx.fillText(`N: ${nLo} → ${nHi}`, 6, baseY);
-    }
-
-    // Center match marker
-    ctx.strokeStyle = PC.centerMark;
-    ctx.lineWidth = 1;
-    ctx.beginPath();
-    ctx.moveTo(cx - 4, cy);
-    ctx.lineTo(cx + 4, cy);
-    ctx.moveTo(cx, cy - 4);
-    ctx.lineTo(cx, cy + 4);
-    ctx.stroke();
-    // multiFeed is captured in the closure; without it in the deps the
-    // chart wouldn't redraw when the descriptor flag flips from its
-    // initial false to the real /examples value (true for bowtie /
-    // hexbeam_5band) — the user saw only one Z* annotation in the
-    // legend because the closure stayed wedged on the single-feed branch.
-  }, [r, x, z0, size, sweep, converge, measured, measFreqMhz, running, convergeRunning, feeds, multiFeed, theme]);
-
-  return <canvas ref={canvasRef} className="smith" />;
-}
-
-// Viewport zoom ceiling. The motivating case (elt_whip, #384) hides 6.35 mm
-// of cage detail inside a 2.44 m extent — a 1:384 scale gap — so the ceiling
-// leaves an order of magnitude of headroom past "inspect the finest catalog
-// detail at canvas size".
-const VIEWPORT_ZOOM_MAX = 10000;
-
-function CurrentCanvas({
-  result,
-  projection,
-  showHeatmap,
-  showEnvelope,
-  showWireLabels,
-  showFeedNames,
-  interactive = false,
-}: {
-  result: SolveResponse | null;
-  projection: Projection;
-  showHeatmap: boolean;
-  showEnvelope: boolean;
-  showWireLabels: boolean;
-  showFeedNames: boolean;
-  // Zoom/pan navigation — main stage only; thumbnails stay inert buttons.
-  interactive?: boolean;
-}) {
-  const theme = useContext(ThemeContext); // repaint on theme toggle (dep below)
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-
-  // Viewport navigation (map-style: cursor-anchored wheel/pinch zoom, drag
-  // pan, double-click/tap or the Fit button to re-fit — survey in PR #384).
-  // zoom=1, pan=0 IS the auto-fit view, so the fit framing keeps tracking
-  // knob drags and re-solves; zoom composes on top as a pure multiplier.
-  // Lives in a ref (mutated at pointer-event rate, drawn via rAF) with a
-  // React mirror of the zoom level for the HUD chip / touch-action gate.
-  const vpRef = useRef({ zoom: 1, panX: 0, panY: 0 });
-  const [vpZoom, setVpZoom] = useState(1);
-  const redrawRef = useRef<() => void>(() => {});
-  const resetViewport = () => {
-    vpRef.current = { zoom: 1, panX: 0, panY: 0 };
-    setVpZoom(1);
-  };
-
-  // The fit frame of the last completed draw (projection + fit centre/scale).
-  // A projection switch carries the viewport over through it — see draw() —
-  // instead of resetting, so a feature under 400× inspection stays under
-  // inspection when the view turns. Cleared on design switch: the previous
-  // design's frame means nothing for the new geometry.
-  const frameRef = useRef<{
-    projection: Projection;
-    hC: number;
-    vC: number;
-    scale: number;
-  } | null>(null);
-
-  // Switching DESIGNS re-fits: the viewport was aimed at the old geometry.
-  // (Projection switches within a design carry the viewport — see draw().)
-  const geometryName = result?.geometry ?? "";
-  useEffect(() => {
-    resetViewport();
-    frameRef.current = null;
-    // The main draw effect below re-runs on any new result, so the re-fit
-    // paints without an explicit redraw.
-  }, [geometryName]);
-
-  useEffect(() => {
-    const canvas = canvasRef.current;
-    if (!canvas) return;
-    const ctx = canvas.getContext("2d");
-    if (!ctx) return;
-
-    const dpr = window.devicePixelRatio || 1;
-    const PC = plotColors();
-    const onResize = () => {
-      const rect = canvas.getBoundingClientRect();
-      canvas.width = Math.floor(rect.width * dpr);
-      canvas.height = Math.floor(rect.height * dpr);
-      ctx.setTransform(dpr, 0, 0, dpr, 0, 0);
-      draw();
-    };
-
-    function draw() {
-      if (!canvas) return;
-      const w = canvas.clientWidth;
-      const h = canvas.clientHeight;
-      ctx!.clearRect(0, 0, w, h);
-
-      // Vertical axis guide.
-      ctx!.strokeStyle = PC.axisFaint;
-      ctx!.lineWidth = 1;
-      ctx!.beginPath();
-      ctx!.moveTo(w / 2, 20);
-      ctx!.lineTo(w / 2, h - 20);
-      ctx!.stroke();
-
-      if (!result) return;
-
-      // Scale anchored to design wavelength. Worst-case extents (in λ):
-      //   horizontal: hf_max × λ/2 ≈ 0.6 λ  (both V and Yagi)
-      //   vertical:   max(V droop, Yagi spacing) ≈ 0.5 λ
-      //
-      // `s` proportionally shrinks every fixed pixel constant (padding,
-      // strokes, envelope amplitude, label sizes) so the rendering looks
-      // the same at thumbnail and main sizes. Floor keeps thumbnails
-      // legible; cap prevents very-large canvases from over-inflating.
-      const refSize = 600;
-      const s = Math.max(0.3, Math.min(1.4, Math.min(w, h) / refSize));
-      const lambdaDesign = result.lambda_design_m;
-      const pad = 50 * s;
-      const barReserveBottom = 40 * s;
-      const FILL = 0.85;
-
-      // Camera projection: an orthonormal screen basis (see PROJECTIONS) —
-      // world point p lands at canvas (h·p, v·p); the camera ray is h×v.
-      // App.tsx sets a per-geometry default (V/fan_dipole → "yz" side,
-      // Yagi/moxon/hexbeam → "xy" top) but the user can override via the
-      // projection toggle in the stage.
-      const projSpec = PROJECTIONS.find((p) => p.id === projection)!;
-      const hVec = projSpec.h;
-      const vVec = projSpec.v;
-      // True for the two elevation views (screen-up IS world z), which gates
-      // the ground reference line below; the isometric's tilted up-vector
-      // draws no line (z=0 doesn't project to a horizontal line there).
-      const upIsZ = vVec[0] === 0 && vVec[1] === 0 && vVec[2] === 1;
-      let hMin = Infinity, hMax = -Infinity;
-      let vMin = Infinity, vMax = -Infinity;
-      // Axis-aligned world bbox too — the projection-switch carry-over needs
-      // a depth estimate along the old camera ray (see below).
-      const bbMin = [Infinity, Infinity, Infinity];
-      const bbMax = [-Infinity, -Infinity, -Infinity];
-      for (const wire of result.wires) {
-        for (const p of wire.knot_positions) {
-          const ph = dot3(hVec, p);
-          const pv = dot3(vVec, p);
-          if (ph < hMin) hMin = ph;
-          if (ph > hMax) hMax = ph;
-          if (pv < vMin) vMin = pv;
-          if (pv > vMax) vMax = pv;
-          for (let a = 0; a < 3; a++) {
-            if (p[a] < bbMin[a]) bbMin[a] = p[a];
-            if (p[a] > bbMax[a]) bbMax[a] = p[a];
-          }
-        }
-      }
-
-      // When ground is enabled and screen-up is world z, expand the visible
-      // vertical range to include z=0 so the ground reference line lands
-      // inside the canvas. Without this, antennas sitting well above the
-      // plane push the ground line off-screen.
-      let vEffMin = vMin, vEffMax = vMax;
-      if (result.ground && upIsZ) {
-        vEffMin = Math.min(vMin, 0);
-        vEffMax = Math.max(vMax, 0);
-      }
-      // Vertical span used to size the canvas. Floor at the wavelength
-      // worst-case so small antennas don't render comically large; grow with
-      // the ground-adjusted antenna span so high antennas zoom out enough
-      // to fit the ground line.
-      const vSpanEff = Math.max(vEffMax - vEffMin, 0.5 * lambdaDesign);
-      // Horizontal span: same floor-with-actual-extent pattern as vertical.
-      // The 0.6λ floor covers the typical V / Yagi worst case; wider
-      // antennas (EDZ at ~1.5λ, fan-dipole 5-band, ...) grow the span
-      // from their actual hMax-hMin so they fit on canvas.
-      const hSpanEff = Math.max(hMax - hMin, 0.6 * lambdaDesign);
-      const scale = FILL * Math.min(
-        (w - 2 * pad) / hSpanEff,
-        (h - pad - barReserveBottom) / vSpanEff,
-      );
-
-      const hC = (hMin + hMax) / 2;
-      const vC = (vEffMin + vEffMax) / 2;
-      const cx = w / 2;
-      const cy = h / 2;
-      const vp = vpRef.current;
-
-      // Projection switch with an active zoom: carry the viewport over
-      // instead of resetting. Reconstruct the world point at the old canvas
-      // centre — its two screen coordinates from the old frame, its depth
-      // along the old camera ray from the geometry point nearest that
-      // centre ray (the wire actually under inspection; the bbox centre
-      // would misplace anything off-centre in depth, e.g. an apex feed) —
-      // then aim the new frame at that point, preserving the absolute px/m
-      // scale so the feature keeps its on-screen size (each view has its
-      // own fit scale, so the relative zoom factor is rescaled). If the
-      // carried zoom clamps to 1, it degrades to a plain fit. Design
-      // switches still hard-reset (frameRef is cleared by the effect above).
-      const frame = frameRef.current;
-      if (frame && frame.projection !== projection && vp.zoom > 1) {
-        const old = PROJECTIONS.find((p) => p.id === frame.projection)!;
-        const oldZscale = frame.scale * vp.zoom;
-        const hCtr = frame.hC - vp.panX / oldZscale;
-        const vCtr = frame.vC + vp.panY / oldZscale;
-        const n = cross3(old.h, old.v);
-        let depth = dot3(n, [
-          (bbMin[0] + bbMax[0]) / 2,
-          (bbMin[1] + bbMax[1]) / 2,
-          (bbMin[2] + bbMax[2]) / 2,
-        ]);
-        let best = Infinity;
-        for (const wire of result.wires) {
-          for (const p of wire.knot_positions) {
-            const dh = dot3(old.h, p) - hCtr;
-            const dv = dot3(old.v, p) - vCtr;
-            const d2 = dh * dh + dv * dv;
-            if (d2 < best) {
-              best = d2;
-              depth = dot3(n, p);
-            }
-          }
-        }
-        const P: Vec3 = [
-          old.h[0] * hCtr + old.v[0] * vCtr + n[0] * depth,
-          old.h[1] * hCtr + old.v[1] * vCtr + n[1] * depth,
-          old.h[2] * hCtr + old.v[2] * vCtr + n[2] * depth,
-        ];
-        const zoomNew = Math.min(
-          VIEWPORT_ZOOM_MAX,
-          Math.max(1, (vp.zoom * frame.scale) / scale),
-        );
-        vp.zoom = zoomNew;
-        if (zoomNew > 1) {
-          const zs = scale * zoomNew;
-          vp.panX = (hC - dot3(hVec, P)) * zs;
-          vp.panY = (dot3(vVec, P) - vC) * zs;
-        } else {
-          vp.panX = 0;
-          vp.panY = 0;
-        }
-        setVpZoom(zoomNew);
-      }
-      frameRef.current = { projection, hC, vC, scale };
-
-      // Compose the user viewport on top of the fit framing. Only geometry
-      // goes through `zscale`; pixel-sized glyphs (strokes, labels, envelope
-      // amplitude, feed dot) stay on `s`, so zooming magnifies the antenna
-      // without ballooning its annotations.
-      const zscale = scale * vp.zoom;
-      const project = (p: [number, number, number]) => ({
-        x: cx + (dot3(hVec, p) - hC) * zscale + vp.panX,
-        y: cy + (vC - dot3(vVec, p)) * zscale + vp.panY, // higher vert value = higher on screen
-      });
-
-      // Ground reference line at world z=0, drawn only on the elevation
-      // views (screen-up is world z) when the backend has ground enabled.
-      // Cosmetic — the math is correct regardless; this just removes the
-      // "where is the ground" guessing game from the side view. vC was
-      // adjusted above to keep this on-canvas, so no bounds check needed
-      // here.
-      if (result.ground && upIsZ) {
-        const groundY = cy + vC * zscale + vp.panY;
-        ctx!.strokeStyle = `rgba(${PC.groundRgb}, 0.55)`;
-        ctx!.lineWidth = 1;
-        ctx!.setLineDash([6, 4]);
-        ctx!.beginPath();
-        ctx!.moveTo(0, groundY);
-        ctx!.lineTo(w, groundY);
-        ctx!.stroke();
-        ctx!.setLineDash([]);
-        ctx!.fillStyle = `rgba(${PC.groundRgb}, 0.85)`;
-        ctx!.font = `${Math.max(8, Math.round(10 * s))}px ui-monospace, monospace`;
-        ctx!.fillText("ground (z = 0)", 8 * s, groundY - 4 * s);
-      }
-
-      // Global current magnitude — use sample arrays when available so the
-      // shared color scale catches mid-segment peaks (B-spline d=2 quadratic
-      // curvature, sinusoidal three-term, B-spline enrichment dip). Falls
-      // back to knot arrays for backends that don't ship samples (PyNEC).
-      let magMaxGlobal = 1e-30;
-      const perWirePts: [number, number, number][][] = [];
-      const perWireMags: number[][] = [];
-      for (const wire of result.wires) {
-        const pts = wire.sample_positions ?? wire.knot_positions;
-        const cre = wire.sample_currents_re ?? wire.knot_currents_re;
-        const cim = wire.sample_currents_im ?? wire.knot_currents_im;
-        const m = cre.map((r, i) => Math.hypot(r, cim[i]));
-        perWirePts.push(pts);
-        perWireMags.push(m);
-        for (const v of m) if (v > magMaxGlobal) magMaxGlobal = v;
-      }
-
-      ctx!.lineCap = "round";
-      ctx!.lineJoin = "round";
-
-      // One wire at a time: wire stroke + envelope.
-      const envScale = 60 * s;
-      const labelFontPx = Math.max(8, Math.round(11 * s));
-      const feedFontPx = Math.max(8, Math.round(12 * s));
-      const feedWireIdx = result.feed_wire_index;
-      for (let wi = 0; wi < result.wires.length; wi++) {
-        const wire = result.wires[wi];
-        const pts = perWirePts[wi];
-        const mags = perWireMags[wi];
-
-        for (let i = 0; i < pts.length - 1; i++) {
-          const a = project(pts[i]);
-          const b = project(pts[i + 1]);
-          if (showHeatmap) {
-            const m = (0.5 * (mags[i] + mags[i + 1])) / magMaxGlobal;
-            ctx!.strokeStyle = currentColor(m);
-            ctx!.lineWidth = (2 + 6 * m) * s;
-          } else {
-            // Plain wires: uniform color/width, no current-magnitude modulation.
-            ctx!.strokeStyle = PC.labelBright;
-            ctx!.lineWidth = 2 * s;
-          }
-          ctx!.beginPath();
-          ctx!.moveTo(a.x, a.y);
-          ctx!.lineTo(b.x, b.y);
-          ctx!.stroke();
-        }
-
-        // Current-waveform envelope: if this is the feed wire (and the feed
-        // isn't at an end), split at the feed knot so a V's per-arm tangent
-        // flip is respected. Otherwise draw one continuous envelope.
-        // feed_knot_index lives in knot-array space; in sample space (knots
-        // interleaved with midpoints) it maps to 2*feed_knot_index.
-        if (showEnvelope) {
-          ctx!.strokeStyle = `rgba(${PC.envelopeRgb}, 0.7)`;
-          ctx!.lineWidth = 1.5 * s;
-          const lastIdx = pts.length - 1;
-          const hasSamples = wire.sample_positions != null;
-          const feedIdx = result.feed_knot_index * (hasSamples ? 2 : 1);
-          if (wi === feedWireIdx && feedIdx > 0 && feedIdx < lastIdx) {
-            drawArmEnvelope(ctx!, pts, mags, magMaxGlobal, project, 0, feedIdx, envScale);
-            drawArmEnvelope(ctx!, pts, mags, magMaxGlobal, project, feedIdx, lastIdx, envScale);
-          } else {
-            drawArmEnvelope(ctx!, pts, mags, magMaxGlobal, project, 0, lastIdx, envScale);
-          }
-        }
-
-        // Wire label near the leftmost knot for multi-wire geometries.
-        if (showWireLabels && result.wires.length > 1) {
-          const lp = project(wire.knot_positions[0]);
-          ctx!.fillStyle = PC.label;
-          ctx!.font = `${labelFontPx}px ui-monospace, monospace`;
-          ctx!.fillText(wire.label, lp.x - 8 * s - ctx!.measureText(wire.label).width, lp.y + 3 * s);
-        }
-      }
-
-      // Feed marker(s). Three sources, in priority order:
-      //  1. `feed_positions[]` (issue #571): one physical feed point per
-      //     declared feed port — a lazy-H fed in phase through a harness has
-      //     two, one per element centre, even though its drive comes from a
-      //     single build_network() source. Rendered with the port name.
-      //  2. `feeds[]` (bowtie 1×2 array): per-feed Z+V, labelled by phase.
-      //  3. the legacy single feed_position / feed_wire_index path.
-      if (result.feed_positions && result.feed_positions.length > 0) {
-        const fps = result.feed_positions;
-        for (let fi = 0; fi < fps.length; fi++) {
-          const feed = project(fps[fi].position);
-          ctx!.fillStyle = PC.feed;
-          ctx!.beginPath();
-          ctx!.arc(feed.x, feed.y, 5 * s, 0, Math.PI * 2);
-          ctx!.fill();
-          if (showFeedNames) {
-            ctx!.font = `${feedFontPx}px ui-monospace, monospace`;
-            const label = fps.length > 1 ? fps[fi].name : "feed";
-            ctx!.fillText(label, feed.x + 8 * s, feed.y - 8 * s);
-          }
-        }
-      } else {
-        const feedList = result.feeds && result.feeds.length > 0
-          ? result.feeds
-          : [{
-              wire_index: feedWireIdx,
-              knot_index: result.feed_knot_index,
-              feed_position: result.feed_position,
-              v_re: 1, v_im: 0,
-              z_re: result.z_in_re, z_im: result.z_in_im,
-            }];
-        for (let fi = 0; fi < feedList.length; fi++) {
-          const f = feedList[fi];
-          const w_ = result.wires[f.wire_index];
-          // Prefer the exact feed point; fall back to the nearest knot.
-          const pos3d = f.feed_position ?? (w_ ? w_.knot_positions[f.knot_index] : undefined);
-          if (!pos3d) continue;
-          const feed = project(pos3d);
-          ctx!.fillStyle = PC.feed;
-          ctx!.beginPath();
-          ctx!.arc(feed.x, feed.y, 5 * s, 0, Math.PI * 2);
-          ctx!.fill();
-          if (showFeedNames) {
-            ctx!.font = `${feedFontPx}px ui-monospace, monospace`;
-            const label = feedList.length > 1
-              ? `feed ${fi} ∠${Math.round(Math.atan2(f.v_im, f.v_re) * 180 / Math.PI)}°`
-              : "feed";
-            ctx!.fillText(label, feed.x + 8 * s, feed.y - 8 * s);
-          }
-        }
-      }
-
-      // Scale bar, centered horizontally under the antenna. At fit zoom it
-      // is the familiar λ/4 bar; once zoomed, λ/4 no longer fits on screen,
-      // so it becomes a map-style bar: a nice round length (1/2/5 × 10^k m)
-      // near a quarter of the canvas width, always true to `zscale`.
-      let barWorld = lambdaDesign / 4;
-      let barLabel = `λ/4 = ${(lambdaDesign / 4).toFixed(2)} m`;
-      if (vp.zoom !== 1) {
-        const target = (0.25 * w) / zscale;
-        const pow = Math.pow(10, Math.floor(Math.log10(target)));
-        const mant = target / pow;
-        barWorld = (mant >= 5 ? 5 : mant >= 2 ? 2 : 1) * pow;
-        barLabel = formatMetres(barWorld);
-      }
-      const barLenPx = barWorld * zscale;
-      const barX0 = (w - barLenPx) / 2;
-      const barY = h - 24 * s;
-      ctx!.strokeStyle = PC.label;
-      ctx!.lineWidth = 1;
-      ctx!.beginPath();
-      ctx!.moveTo(barX0, barY);
-      ctx!.lineTo(barX0 + barLenPx, barY);
-      ctx!.moveTo(barX0, barY - 4 * s);
-      ctx!.lineTo(barX0, barY + 4 * s);
-      ctx!.moveTo(barX0 + barLenPx, barY - 4 * s);
-      ctx!.lineTo(barX0 + barLenPx, barY + 4 * s);
-      ctx!.stroke();
-      ctx!.fillStyle = PC.labelBright;
-      ctx!.font = `${labelFontPx}px ui-monospace, monospace`;
-      const labelW = ctx!.measureText(barLabel).width;
-      ctx!.fillText(barLabel, (w - labelW) / 2, barY - 8 * s);
-    }
-
-    onResize();
-    const obs = new ResizeObserver(onResize);
-    obs.observe(canvas);
-    redrawRef.current = draw;
-    if (!interactive) return () => obs.disconnect();
-
-    // ---- viewport navigation ------------------------------------------
-    // Draws coalesce to one per frame: pointer/wheel events mutate vpRef
-    // and schedule a rAF repaint.
-    let raf = 0;
-    const scheduleDraw = () => {
-      if (raf) return;
-      raf = requestAnimationFrame(() => {
-        raf = 0;
-        draw();
-      });
-    };
-
-    // Zoom by `factor` keeping the canvas point (ax, ay) fixed — the world
-    // point under the cursor/pinch centre stays under it. Zooming all the
-    // way back out lands exactly on the fit view (pan snaps to 0).
-    const applyZoom = (factor: number, ax: number, ay: number) => {
-      const v = vpRef.current;
-      const z = Math.min(VIEWPORT_ZOOM_MAX, Math.max(1, v.zoom * factor));
-      const f = z / v.zoom;
-      const cx = canvas.clientWidth / 2;
-      const cy = canvas.clientHeight / 2;
-      v.panX = ax - cx - (ax - cx - v.panX) * f;
-      v.panY = ay - cy - (ay - cy - v.panY) * f;
-      v.zoom = z;
-      if (z === 1) {
-        v.panX = 0;
-        v.panY = 0;
-      }
-      setVpZoom(z);
-      scheduleDraw();
-    };
-
-    // Wheel zoom, ~1.2× per detent, exponential so trackpads feel smooth.
-    // Native non-passive listener for the same reason as the VFO dial:
-    // React's onWheel can't preventDefault the page scroll.
-    const onWheel = (e: WheelEvent) => {
-      e.preventDefault();
-      const rect = canvas.getBoundingClientRect();
-      const dy = e.deltaMode === 1 ? e.deltaY * 33 : e.deltaY; // line-mode → px
-      applyZoom(Math.exp(-dy * 0.002), e.clientX - rect.left, e.clientY - rect.top);
-    };
-
-    // Pointer state: one pointer drags (pan — only when zoomed, so at fit a
-    // touch drag stays with the mobile carousel swipe), two pinch-zoom.
-    const pointers = new Map<number, { x: number; y: number; downX: number; downY: number }>();
-    let moved = false;
-    let lastTap = { t: 0, x: 0, y: 0 };
-    const posOf = (e: PointerEvent) => {
-      const r = canvas.getBoundingClientRect();
-      return { x: e.clientX - r.left, y: e.clientY - r.top };
-    };
-
-    const onPointerDown = (e: PointerEvent) => {
-      if (e.pointerType === "mouse" && e.button !== 0 && e.button !== 1) return;
-      try {
-        canvas.setPointerCapture(e.pointerId);
-      } catch {
-        // Capture is best-effort: a drag that leaves the canvas just ends.
-      }
-      const p = posOf(e);
-      pointers.set(e.pointerId, { ...p, downX: p.x, downY: p.y });
-      moved = false;
-      if (e.pointerType === "mouse") e.preventDefault(); // middle-click autoscroll
-    };
-
-    const onPointerMove = (e: PointerEvent) => {
-      const prev = pointers.get(e.pointerId);
-      if (!prev) return;
-      const p = posOf(e);
-      if (Math.hypot(p.x - prev.downX, p.y - prev.downY) > 4) moved = true;
-      if (pointers.size === 2) {
-        // Pinch: translate by the midpoint delta, zoom by the distance
-        // ratio about the new midpoint.
-        const other = [...pointers.entries()].find(([id]) => id !== e.pointerId)![1];
-        const oldMid = { x: (prev.x + other.x) / 2, y: (prev.y + other.y) / 2 };
-        const oldDist = Math.hypot(prev.x - other.x, prev.y - other.y) || 1;
-        const newMid = { x: (p.x + other.x) / 2, y: (p.y + other.y) / 2 };
-        const newDist = Math.hypot(p.x - other.x, p.y - other.y) || 1;
-        const v = vpRef.current;
-        v.panX += newMid.x - oldMid.x;
-        v.panY += newMid.y - oldMid.y;
-        applyZoom(newDist / oldDist, newMid.x, newMid.y);
-      } else if (pointers.size === 1 && vpRef.current.zoom > 1) {
-        const v = vpRef.current;
-        v.panX += p.x - prev.x;
-        v.panY += p.y - prev.y;
-        scheduleDraw();
-      }
-      pointers.set(e.pointerId, { ...prev, x: p.x, y: p.y });
-    };
-
-    const onPointerUp = (e: PointerEvent) => {
-      const had = pointers.delete(e.pointerId);
-      try {
-        canvas.releasePointerCapture(e.pointerId);
-      } catch {
-        // Never captured (see above) — nothing to release.
-      }
-      if (!had || moved || e.type === "pointercancel") return;
-      // Clean tap/click: double within 350 ms & 30 px re-fits.
-      const now = performance.now();
-      const p = posOf(e);
-      if (now - lastTap.t < 350 && Math.hypot(p.x - lastTap.x, p.y - lastTap.y) < 30) {
-        resetViewport();
-        scheduleDraw();
-        lastTap = { t: 0, x: 0, y: 0 };
-      } else {
-        lastTap = { t: now, x: p.x, y: p.y };
-      }
-    };
-
-    canvas.addEventListener("wheel", onWheel, { passive: false });
-    canvas.addEventListener("pointerdown", onPointerDown);
-    canvas.addEventListener("pointermove", onPointerMove);
-    canvas.addEventListener("pointerup", onPointerUp);
-    canvas.addEventListener("pointercancel", onPointerUp);
-    return () => {
-      obs.disconnect();
-      if (raf) cancelAnimationFrame(raf);
-      canvas.removeEventListener("wheel", onWheel);
-      canvas.removeEventListener("pointerdown", onPointerDown);
-      canvas.removeEventListener("pointermove", onPointerMove);
-      canvas.removeEventListener("pointerup", onPointerUp);
-      canvas.removeEventListener("pointercancel", onPointerUp);
-    };
-  }, [result, projection, showHeatmap, showEnvelope, showWireLabels, showFeedNames, theme, interactive]);
-
-  const zoomed = vpZoom > 1.001;
-  return (
-    <div className="canvas-viewport">
-      <canvas
-        ref={canvasRef}
-        style={
-          interactive
-            ? {
-                // At fit, touch drags belong to the page (mobile carousel
-                // swipe); pinch is never a browser gesture here, so a
-                // two-finger zoom always reaches us. Once zoomed, the canvas
-                // owns all touches for panning until re-fit.
-                touchAction: zoomed ? "none" : "pan-x pan-y",
-                cursor: zoomed ? "grab" : "zoom-in",
-              }
-            : undefined
-        }
-      />
-      {interactive && (
-        <div className="viewport-hud">
-          {zoomed && (
-            <span className="viewport-zoom">
-              {vpZoom >= 10 ? Math.round(vpZoom) : vpZoom.toFixed(1)}×
-            </span>
-          )}
-          <button
-            className="viewport-fit"
-            disabled={!zoomed}
-            onClick={() => {
-              resetViewport();
-              redrawRef.current();
-            }}
-            title="Zoom to fit (or double-click the canvas)"
-          >
-            Fit
-          </button>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function drawArmEnvelope(
-  ctx: CanvasRenderingContext2D,
-  knots: [number, number, number][],
-  mags: number[],
-  magMax: number,
-  project: (p: [number, number, number]) => { x: number; y: number },
-  start: number,
-  end: number,
-  envScale: number,
-) {
-  if (end <= start) return;
-
-  // Per-segment normal in canvas space, oriented toward screen-up so V-style
-  // arms put their envelopes "above" the wire. For axis-aligned vertical
-  // segments ny is exactly zero and the flip is a no-op; that's fine — what
-  // matters is that the moxon's adjacent perpendicular segments get
-  // *different* normals so the bend-break below catches the corner.
-  const segN: { nx: number; ny: number }[] = [];
-  for (let i = start; i < end; i++) {
-    const p = project(knots[i]);
-    const q = project(knots[i + 1]);
-    const dx = q.x - p.x;
-    const dy = q.y - p.y;
-    const len = Math.hypot(dx, dy) || 1;
-    let nx = -dy / len;
-    let ny = dx / len;
-    if (ny > 0) {
-      nx = -nx;
-      ny = -ny;
-    }
-    segN.push({ nx, ny });
-  }
-
-  // Walk runs of segments whose normals agree (within ~3°), and start a new
-  // sub-path at each bend. Without this, a connected envelope at a 90°
-  // corner zigzags across the corner since the two adjacent segments offset
-  // their knots in perpendicular directions.
-  const BEND_TOL = 0.9986;  // cos(3°)
-  ctx.beginPath();
-  let s = 0;
-  while (s < segN.length) {
-    let e = s;
-    while (
-      e + 1 < segN.length &&
-      segN[e].nx * segN[e + 1].nx + segN[e].ny * segN[e + 1].ny >= BEND_TOL
-    ) {
-      e++;
-    }
-    const { nx, ny } = segN[s];
-    for (let k = s; k <= e + 1; k++) {
-      const ki = start + k;
-      const p = project(knots[ki]);
-      const offset = (mags[ki] / magMax) * envScale;
-      const ex = p.x + nx * offset;
-      const ey = p.y + ny * offset;
-      if (k === s) ctx.moveTo(ex, ey);
-      else ctx.lineTo(ex, ey);
-    }
-    s = e + 1;
-  }
-  ctx.stroke();
-}
-
-// Plot colors are pulled from CSS custom properties so the <canvas>
-// views theme from the same tokens as the DOM chrome. Fallbacks
-// reproduce the original dark palette, so missing vars are harmless.
-function plotColors() {
-  const cs = getComputedStyle(document.documentElement);
-  const v = (name: string, fb: string): string => {
-    const val = cs.getPropertyValue(name).trim();
-    return val || fb;
-  };
-  return {
-    bg: v("--plot-bg", "#0d1015"),
-    bgRgb: v("--plot-bg-rgb", "13, 16, 21"),
-    grid: v("--plot-grid", "#2a313d"),
-    axis: v("--plot-axis", "#3a4150"),
-    axisFaint: v("--plot-axis-faint", "#23272f"),
-    labelDim: v("--plot-label-dim", "#4a5160"),
-    label: v("--plot-label", "#7b8493"),
-    labelBright: v("--plot-label-bright", "#9aa3b2"),
-    labelStrong: v("--plot-label-strong", "#cdd5e0"),
-    centerMark: v("--plot-center-mark", "#5a6170"),
-    spoke: v("--plot-spoke", "rgba(180, 140, 250, 0.7)"),
-    lobeRgb: v("--plot-lobe-rgb", "255, 209, 102"),
-    necRgb: v("--plot-nec-rgb", "110, 220, 255"),
-    groundRgb: v("--plot-ground-rgb", "140, 110, 70"),
-    envelopeRgb: v("--plot-envelope-rgb", "118, 208, 255"),
-    feed: v("--plot-feed", "#ffd166"),
-    // Measured-overlay locus (issue #595). Violet: unused elsewhere on the
-    // Smith chart, so it never reads as another feed.
-    measured: v("--plot-measured", "#b48cfa"),
-  };
-}
-
-// Current-magnitude heatmap ramp, also CSS-driven. Read once and cached
-// (currentColor runs per wire segment, so getComputedStyle must not be
-// called in the loop). Fallbacks are the original cool->warm stops.
-let _currentRampCache: [number, [number, number, number]][] | null = null;
-function currentRamp(): [number, [number, number, number]][] {
-  if (_currentRampCache) return _currentRampCache;
-  const cs = getComputedStyle(document.documentElement);
-  const tri = (name: string, fb: [number, number, number]): [number, number, number] => {
-    const s = cs.getPropertyValue(name).trim();
-    if (!s) return fb;
-    const p = s.split(",").map((n) => parseInt(n.trim(), 10));
-    return p.length === 3 && p.every((n) => !Number.isNaN(n)) ? [p[0], p[1], p[2]] : fb;
-  };
-  _currentRampCache = [
-    [0.0, tri("--plot-current-0", [40, 64, 96])],
-    [0.25, tri("--plot-current-1", [60, 140, 200])],
-    [0.5, tri("--plot-current-2", [118, 208, 255])],
-    [0.75, tri("--plot-current-3", [255, 209, 102])],
-    [1.0, tri("--plot-current-4", [255, 130, 80])],
-  ];
-  return _currentRampCache;
-}
-
-function currentColor(t: number): string {
-  // Cool → warm ramp: dim blue → cyan → yellow → orange.
-  const stops = currentRamp();
-  for (let i = 1; i < stops.length; i++) {
-    const [t0, c0] = stops[i - 1];
-    const [t1, c1] = stops[i];
-    if (t <= t1) {
-      const f = (t - t0) / (t1 - t0 || 1);
-      const r = Math.round(c0[0] + (c1[0] - c0[0]) * f);
-      const g = Math.round(c0[1] + (c1[1] - c0[1]) * f);
-      const b = Math.round(c0[2] + (c1[2] - c0[2]) * f);
-      return `rgb(${r},${g},${b})`;
-    }
-  }
-  return "rgb(255,130,80)";
 }
